feat: added KanbanCard component with select props; added click card route feature leading to KanBanCard page; added KanbanCard[id] page file; refactor: organized components folder;
