+++ capstone-project/src/app/components/ClerkAuthenticator/MainContent/page.tsx
