+++ capstone-project/src/app/components/KanbanBoard/KanbanColumn.tsx
@@ -1,25 +1,33 @@
 import React from "react";
+import KanbanCard from "./KanbanCard";
+import { useRouter } from "next/router";
 
 type KanbanColumnProps = {
     title: string;
-    cards: [id: string, title: string];
+    cards:{id: string, title: string}[];
 };
-
-//add 
-
-const handleClick = () => {
-//route to KanbanCard.tsx page
-  };
   
 const KanbanColumn: React.FC<KanbanColumnProps> = ({ title, cards }) => {
+    const router = useRouter();
+
+    const handleClick = (id: string) => {
+        router.push(`/KanbanCard/${id}`);
+    };
+
     return (
-        <div>
-            <h1>
-                {title}
-            </h1>
-            <button onClick={handleClick}>
-                {cards}
-            </button>
+        <div className="">
+            <h2 className="">{title}</h2>
+            <div className="">
+                {cards.map((card) => (
+                    <div
+                        key={card.id}
+                        className="cursor-pointer"
+                        onClick={() => handleClick(card.id)}
+                    >
+                        <KanbanCard id={card.id} title={card.title} description={""} tags={""} dueDate={""} />
+                    </div>
+                ))}
+            </div>
         </div>
     );
 };
