+++ capstone-project/src/app/components/Rive/RiveNoSSR/RiveNoSSR.tsx
