+++ capstone-project/src/app/pages/KanbanCard/[id].tsx
@@ -1,0 +1,4 @@
+import { useRouter } from "next/router";
+import React from "react";
+
+export default KanbanCardPage;
